# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "VNERP API",
    "version": "1.0.0",
    "description": "ERP System API for Printing Industry"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "summary": "No description",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
...
  }
}